services/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
geoalchemy2==0.15.2
watchfiles==0.24.0
